+++ CHANGELOG.md
@@ -9,3 +9,12 @@ Updated LICENSE release.
 ## 0.0.3
 
 Updated READEME file for user understanding.
+
+## 0.0.4
+
+added customizations 
+hour hand1 hand2,
+minute hand1 hand2,
+second hand1 hand2,
+
+